requirements.txt
flask>=2.2.5
tensorflow>=2.12.0
numpy>=1.24.3
pandas>=1.5.3
mysql-connector-python>=8.0.33
scikit-learn>=1.2.2
matplotlib>=3.7.1
Pillow>=9.5.0
